fold AbstractStateManager into the AE class hierarchy
Drop the standalone AbstractStateManager.h/.cpp.  The dense base now
lives next to AbstractInterpretation; the semi-sparse and full-sparse
managers sit beside SparseAbstractInterpretation:

  AbstractStateManager                    (dense, AI files)
    └── SemiSparseAbstractStateManager      (SparseAI files)
          └── FullSparseAbstractStateManager (adds SVFG)

Each AE subclass picks its manager via virtual createStateMgr().  The
per-cycle ValVar precompute is gated by a virtual needsCycleValVars(),
so dense and full-sparse skip it.  Together this removes every
Options::AESparsity() check that lived inside the manager and inside
PreAnalysis::initCycleValVars; mode dispatch is now polymorphic.
